requirements.txt
ffmpeg-python>=0.2.0
openai-whisper>=20231117

